[internal/party-server-ish] Make run primarily on postgres
As edge-rest and party-server-ish must use the same database,
swap to postgres for local development and do basic testing.
