--- adks/admin-csharp/csharp-admin-client/.gitignore
@@ -1,12 +1,0 @@
-
-target
-.idea
-*.iml
-git_push.sh
-appveyor.yml
-src
-bin
-obj
-*.sln.DotSettings.user
-.openapi-generator
-
