--- adks/admin-csharp/csharp-admin-client/Dockerfile-build
@@ -1,9 +1,0 @@
-FROM mcr.microsoft.com/dotnet/sdk:3.1 AS build-env
-ARG Version
-WORKDIR /src
-ADD . /src
-ADD AdminSDKTestProject /src
-RUN dotnet restore IO.FeatureHub.MR.sln
-RUN dotnet publish IO.FeatureHub.MR.sln -c Release -o out
-RUN dotnet build IO.FeatureHub.MR.sln /p:PackageVersion=$Version -c Release --no-restore
-
