--- adks/admin-csharp/csharp-admin-client/IO.FeatureHub.MR.sln
@@ -1,31 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IO.FeatureHub.MR", "src\IO.FeatureHub.MR\IO.FeatureHub.MR.csproj", "{70E60832-BF91-4A60-9697-E3E763B60DDB}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdminSDKTestProject", "AdminSDKTestProject\AdminSDKTestProject.csproj", "{4C2D384D-9E45-40B8-BFD8-F98562603E43}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{70E60832-BF91-4A60-9697-E3E763B60DDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{70E60832-BF91-4A60-9697-E3E763B60DDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{70E60832-BF91-4A60-9697-E3E763B60DDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{70E60832-BF91-4A60-9697-E3E763B60DDB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{19F1DEBC-DE5E-4517-8062-F000CD499087}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{19F1DEBC-DE5E-4517-8062-F000CD499087}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{19F1DEBC-DE5E-4517-8062-F000CD499087}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{19F1DEBC-DE5E-4517-8062-F000CD499087}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4C2D384D-9E45-40B8-BFD8-F98562603E43}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C2D384D-9E45-40B8-BFD8-F98562603E43}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C2D384D-9E45-40B8-BFD8-F98562603E43}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C2D384D-9E45-40B8-BFD8-F98562603E43}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
