--- adks/admin-csharp/csharp-admin-client/README.md
@@ -1,3 +1,0 @@
-# DOTNET Admin SDK
-
-
