--- adks/admin-csharp/csharp-admin-client/REAL_README.md
@@ -1,1 +1,0 @@
-This is the real readme
