+++ adks/e2e-sdk/Dockerfile
@@ -17,5 +17,5 @@ WORKDIR /app
 
 ADD app /app/app/
 ADD features /app/features/
-COPY package.json package-lock.json tsconfig.json run.sh run-e2e.sh /app/
+COPY package.json package-lock.json tsconfig.json run.sh /app/
 RUN npm install && npm run link && npm run build
