+++ adks/e2e-sdk/features/examples.feature
@@ -1,3 +1,4 @@
+@allvariants @streamingvariants
 Feature: I setup the example features to be able to run the examples
 
   Background:
