+++ adks/e2e-sdk/features/feature-groups.feature
@@ -1,3 +1,4 @@
+@allvariants @streamingvariants
 Feature: Feature groups should operate as per definition
 
   Background:
