+++ adks/e2e-sdk/features/flags.feature
@@ -1,3 +1,4 @@
+@allvariants @streamingvariants
 Feature: All flag based functionality works as expected
 
   Background:
