+++ adks/e2e-sdk/features/lots_of_features.feature
@@ -1,3 +1,4 @@
+@allvariants @streamingvariants
 Feature: I setup a lot of features
 
   Background:
