+++ adks/e2e-sdk/features/strings.feature
@@ -1,3 +1,4 @@
+@allvariants @streamingvariants
 Feature: All string based functionality works as expected
 
   Background:
