+++ backend/build.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # make sure you have built all artifacts first
 if [ $# -eq 0 ]
   then
@@ -13,5 +13,5 @@ else
   BUILD_PARAMS="$BUILD_PARAMS -Djib.to.tags=latest"
 fi
 DOCKER_PREFIX="${OVERRIDE_DOCKER_PREFIX:-featurehub}"
-mvn -f pom-packages.xml -DskipTests $BUILD_PARAMS -Dapp.baseimage.prefix=docker:// -Ddocker.project.prefix=$DOCKER_PREFIX -Ddocker-cloud-build=true -Dbuild.version=$VERSION clean install
+mvn -f pom-packages.xml -DskipTests $BUILD_PARAMS -Dapp.baseimage.prefix=docker:// -Ddocker.project.prefix=$DOCKER_PREFIX -Ddocker-cloud-build=true -Dbuild.version=$VERSION clean install && git add ../infra/api-bucket/files
