+++ backend/edge-rest/pom.xml
@@ -65,6 +65,11 @@
     </dependency>
     <dependency>
       <groupId>io.featurehub.mr</groupId>
+      <artifactId>mr-db-services</artifactId>
+      <version>1.1-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>io.featurehub.mr</groupId>
       <artifactId>mr-db-models</artifactId>
       <version>1.1-SNAPSHOT</version>
     </dependency>
