+++ backend/edge-rest/src/test/resources/application.properties
@@ -1,6 +1,6 @@
 jersey.cors.headers=X-Requested-With,Authorization,Content-type,Accept-Version,Content-MD5,CSRF-Token,cache-control,x-featurehub,Baggage
-db.password=
-db.username=sa
-db.driver=org.h2.Driver
+db.url=jdbc:postgresql://localhost:5432/featurehub
+db.username=featurehub
+db.password=featurehub
 db.connections=3
 server.port=8064
