+++ backend/mr-db-services/src/main/kotlin/io/featurehub/db/publish/CacheSourceFeatureGroupApi.kt
@@ -4,6 +4,7 @@ import io.featurehub.mr.model.RolloutStrategy
 import java.util.*
 
 interface CacheSourceFeatureGroupApi {
+  // returns a Map<Feature-ID,List-of-Rollout-Strategies> (note NOT feature-value id)
   fun collectStrategiesFromGroupsForEnvironment(envId: UUID): Map<UUID, List<RolloutStrategy>>
   fun collectStrategiesFromGroupsForEnvironmentFeature(envId: UUID, featureId: UUID): List<RolloutStrategy>
 }
