+++ backend/mr-db-sql/docker-databases/postgres/docker-compose.yaml
@@ -2,7 +2,7 @@ version: '3.1'
 
 services:
   db:
-    image: postgres:12-alpine
+    image: postgres:14-alpine
     restart: always
     environment:
       POSTGRES_PASSWORD: mypassword
