+++ backend/party-server-ish/src/test/resources/application.properties
@@ -1,11 +1,10 @@
 server.gracePeriodInSeconds=20
-db.url=jdbc:h2:mem:tests
-db.password=
-db.username=sa
-db.driver=org.h2.Driver
+db.url=jdbc:postgresql://localhost:5432/featurehub
+db.username=featurehub
+db.password=featurehub
 register.url=http://localhost:8903/register-url?token=%s
 jersey.cors.headers=X-Requested-With,Authorization,Content-type,Accept-Version,Content-MD5,CSRF-Token,x-ijt,cache-control,x-featurehub,Baggage
 portfolio.admin.group.suffix=Administrators
-#server.port=88
+server.port=8085
 #run.nginx=true
