+++ build_all_and_test.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-cd backend && mvn -f pom-first.xml install && cd .. && mvn -T4C clean install
+cd backend && mvn -f pom-first.xml install && cd .. && mvn -T4C clean install && git add infra/api-bucket/files
